Controller's post Task<long> ==> Task<int>

+++ Software/host/Controllers/SchemesController.cs
@@ -23,9 +23,9 @@ namespace host.Controllers
 
         
         [HttpPost]
-        public Task<long> Post([FromBody]CreateSchemeCommand value)
+        public Task<int> Post([FromBody]CreateSchemeCommand value)
         {
-            return CommandDispatcher.DispatchAsync<CreateSchemeCommand, long>(value);
+            return CommandDispatcher.DispatchAsync<CreateSchemeCommand, int>(value);
         }
